fix: invoice 500 error

+++ database/migrations/2024_05_02_063857_create_billinginfo_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->unsignedBigInteger('event_id');
             $table->foreign('event_id')->references('id')->on('meetings');
             $table->float('bill_amount');
-            $table->float('deposits');
-            $table->float('paymentCredit')->nullable();
+            $table->double('deposits');
+            $table->double('paymentCredit')->nullable();
             $table->float('adjustments');
             $table->float('latefee');
             $table->float('other');
